split prebuilt rules install and upgrade error handling tests
